+++ src/view/main.html
@@ -1,0 +1,120 @@
+<!-- main.html -->
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Ocean Protocol Extension</title>
+    <style>
+      ${styles}
+    </style>
+  </head>
+  <body>
+    <div class="section">
+      <div id="setupHeader" class="section-header active">
+        <span class="chevron">&#9658;</span>Setup
+      </div>
+      <div id="setup" class="section-content active">
+        <div class="container">
+          <label for="rpcUrl">RPC URL</label>
+          <input id="rpcUrl" placeholder="RPC URL" value="http://127.0.0.1:8545" />
+
+          <label for="nodeUrl">Ocean Node URL</label>
+          <input
+            id="nodeUrl"
+            placeholder="Ocean Node URL"
+            value="http://127.0.0.1:8000"
+          />
+
+          <label for="privateKeyInput">Private Key</label>
+          <input
+            id="privateKeyInput"
+            type="password"
+            placeholder="Enter your private key"
+          />
+        </div>
+      </div>
+    </div>
+
+    <div class="section">
+      <div id="getAssetHeader" class="section-header">
+        <span class="chevron">&#9658;</span>Get Asset Details
+      </div>
+      <div id="getAsset" class="section-content">
+        <div class="container">
+          <label for="didInput">Ocean Asset DID</label>
+          <input id="didInput" placeholder="Enter the DID for the asset" />
+          <button id="getAssetDetailsBtn">Get Asset DDO</button>
+        </div>
+      </div>
+    </div>
+
+    <div class="section">
+      <div id="publishHeader" class="section-header">
+        <span class="chevron">&#9658;</span>Publish Asset
+      </div>
+      <div id="publish" class="section-content">
+        <div class="container">
+          <button id="selectFileBtn">Select Asset File</button>
+          <div id="selectedFilePath"></div>
+          <button id="publishAssetBtn">Publish Asset</button>
+        </div>
+      </div>
+    </div>
+
+    <div class="section">
+      <div id="computeHeader" class="section-header">
+        <span class="chevron">&#9658;</span>Start Compute Job
+      </div>
+      <div id="compute" class="section-content">
+        <div class="container">
+          <label>Dataset</label>
+          <button id="selectDatasetBtn">Select Dataset File</button>
+          <div id="selectedDatasetPath" class="selectedFile"></div>
+
+          <label>Algorithm</label>
+          <button id="selectAlgorithmBtn">Select Algorithm File</button>
+          <div id="selectedAlgorithmPath" class="selectedFile"></div>
+
+          <label for="nodeUrlInput">Node URL (including port)</label>
+          <input id="nodeUrlInput" placeholder="Enter compute environment ID" />
+
+          <button id="startComputeBtn">Start Compute Job</button>
+        </div>
+      </div>
+    </div>
+
+    <div class="section">
+      <div id="p2pHeader" class="section-header">
+        <span class="chevron">&#9658;</span>P2P
+      </div>
+      <div id="p2p" class="section-content">
+        <div class="container">
+          <label>Node ID:</label>
+          <div id="nodeIdDisplay">${nodeId}</div>
+          <br />
+          <button id="getOceanPeersBtn">Get Ocean Peers</button>
+        </div>
+      </div>
+    </div>
+
+    <div class="section">
+      <div id="downloadHeader" class="section-header">
+        <span class="chevron">&#9658;</span>Download Asset
+      </div>
+      <div id="download" class="section-content">
+        <div class="container">
+          <label for="assetDidInput">Asset DID</label>
+          <input id="assetDidInput" placeholder="Enter your asset DID" />
+          <label for="pathInput">File Path</label>
+          <input id="pathInput" placeholder="Enter your file path" />
+          <button id="downloadAssetBtn">Download Asset</button>
+        </div>
+      </div>
+    </div>
+
+    <script>
+      ${scripts}
+    </script>
+  </body>
+</html>
